+++ assets/js/blocks/cart-checkout/checkout/inner-blocks/checkout-order-summary-block/edit.tsx
@@ -10,7 +10,7 @@ import { getSetting } from '@woocommerce/settings';
  * Internal dependencies
  */
 import Block from './block';
-import { useBlockPropsWithLocking } from '../../../shared';
+import { useBlockPropsWithLocking } from '../../hacks';
 
 export const Edit = ( {
 	attributes,
