--- assets/js/blocks/cart-checkout/shared/index.tsx
@@ -1,3 +1,0 @@
-export * from './hacks';
-export * from './editor-utils';
-export * from './use-forced-layout';
